solaris fixes

+++ logging.c
@@ -84,7 +84,7 @@ void log_request(struct wsgi_request *wsgi_req) {
 	}
 
 	rlen = snprintf(logpkt, 4096, "[pid: %d|app: %d|req: %d/%llu] %.*s (%.*s) {%d vars in %d bytes} [%.*s] %.*s %.*s => generated %llu bytes in %ld msecs%s(%.*s %d) %d headers in %d bytes (%d async switches on async core %d)\n",
-		uwsgi.mypid,
+		(int) uwsgi.mypid,
 		wsgi_req->app_id,
 		app_req,
 		(unsigned long long ) uwsgi.workers[0].requests,
